+++ packages/server/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @rpgjs/server
 
+## 5.0.0-beta.10
+
+### Patch Changes
+
+- @rpgjs/testing@5.0.0-beta.10
+
 ## 5.0.0-beta.9
 
 ### Major Changes