+++ packages/studio/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @rpgjs/studio
 
+## 5.0.0-beta.10
+
+### Patch Changes
+
+- Updated dependencies
+  - @rpgjs/client@5.0.0-beta.10
+  - @rpgjs/action-battle@5.0.0-beta.10
+  - @rpgjs/server@5.0.0-beta.10
+  - @rpgjs/vite@5.0.0-beta.10
+
 ## 5.0.0-beta.9
 
 ### Major Changes