+++ packages/vite/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @rpgjs/vite
 
+## 5.0.0-beta.10
+
+### Patch Changes
+
+- @rpgjs/server@5.0.0-beta.10
+
 ## 5.0.0-beta.9
 
 ### Major Changes